2/x move ResultSet into a separate module to prevent Import Error
